Introduce variables
Introduce the variables that have not been introduced. Some of the work here is based on guesses of what has been the intention, but none of the changes modify any of the current logic.

The intention is a bit more unclear with variables 'UAMpredBG' and 'COBpredBG' around line 685. It seems the variables have been renamed and should be 'minUAMPredBG' and 'minCOBPredBG' instead. However, making that change would actually change the functionality of the code, so let's verify the intention first.

In javascript, by the way, variables in a function are hoisted to the function scope. Therefore it would be best to declare them all at the beginning of the function, not ad hoc. Even better, of course, would be to use 'let' and 'const' keywords. However, let's refactor in small increments.